Added basic AI movement
Created an EntityBehavior file that contains a basic enum of different entity behaviors, mostly used to Enemies. Modified the Entity.Move function so that we can send an actual number instead of a number between 0 and 1. Added a getHorizontalDistanceToEntity that returns -1, 1, or 0 depending on the location of the entity passed in. Made some variables public that needed to be public.